+++ server/src/events/entity_manipulation.rs
@@ -543,7 +543,16 @@ pub fn handle_delete(server: &mut Server, entity: EcsEntity) {
 pub fn handle_land_on_ground(server: &Server, entity: EcsEntity, vel: Vec3<f32>) {
     let ecs = server.state.ecs();
 
-    if vel.z <= -30.0 {
+    // The second part of this if statement disables all fall damage when in the
+    // water. This was added as a *temporary* fix a bug that causes you to take
+    // fall damage while swimming downwards. FIXME: Fix the actual bug and
+    // remove the following relevant part of the if statement.
+    if vel.z <= -30.0
+        && ecs
+            .read_storage::<PhysicsState>()
+            .get(entity)
+            .map_or(true, |ps| ps.in_liquid().is_none())
+    {
         let char_states = ecs.read_storage::<CharacterState>();
 
         let reduced_vel_z = if let Some(CharacterState::DiveMelee(c)) = char_states.get(entity) {