+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added SFX to the new sword abilities
 - Fixed various issues with showing the correct text hint for interactable blocks.
 - Intert entities like arrows no longer obstruct interacting with nearby entities/blocks.
+- Underwater fall damage
 
 ## [0.14.0] - 2023-01-07